% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_functions.R
\name{calculate_gr_ct_relationships}
\alias{calculate_gr_ct_relationships}
\title{Calculate GR CT relationship for one R0}
\usage{
calculate_gr_ct_relationships(
  model_pars,
  viral_load_pars,
  times,
  lastday = 35,
  cts = seq(0, 40, by = 1),
  population_n = 1,
  convert_daily = TRUE,
  tstep = 1,
  symptom_surveillance = FALSE
)
}
\description{
Calculate GR CT relationship for one R0
}
\examples{
times <- 0:150 # Duration of epidemic in days
lastday <- 35 # For each individual, how long do we assume their measurements can contribute towards the model? 
cts <- seq(0,40,by=1)
model_pars <- c("R0"=R0,"infectious"=5,"incubation"=2,"I0"=0.0001,"t0"=0)
viral_load_pars <- c(beta = 0.1, tshift = 0, desired_mode = 5, viral_peak = 19.7, 
                    obs_sd = 5, sd_mod = 0.789, sd_mod_wane = 14, true_0 = 40, 
                    intercept = 40, LOD = 3, incu = 5, t_switch = 13.3, level_switch = 38, 
                    wane_rate2 = 1000, prob_detect = 0.103, t_unit = 1)
calculate_gr_ct_relationships(model_pars, viral_load_pars, times, lastday, cts)
}
