% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_posterior_func.R
\name{create_post_func_seeirr}
\alias{create_post_func_seeirr}
\title{Creates the posterior function used in the MCMC framework for the SEEIRR model}
\usage{
create_post_func_seeirr(
  parTab,
  data,
  ts,
  INCIDENCE_FUNC = detectable_SEEIRRModel,
  PRIOR_FUNC = NULL,
  ver = "likelihood"
)
}
\arguments{
\item{parTab}{A vector of model parameters with names corresponding to the parameter
control table.}

\item{data}{A tibble containing three columns: date for the sample collection date,
POS for the number of individuals who test positive, and NEG for the number of
individuals who test negative.}

\item{ts}{The times over which the SEEIRR model is solved. Used to subset the modeled
prevalence by the sample collection dates.}

\item{INCIDENCE_FUNC}{A function that returns a vector of daily infection probabilities/
incidence. Set to 'detectable_SEEIRRModel' by default.}

\item{PRIOR_FUNC}{A function for the prior distribution. NULL by default.}

\item{ver}{Can be set to 'model' or 'likelihood.' 'model' returns the prevalence
from the SEEIRR model rather than the posterior probability. Set to 'likelihood' by default.}
}
\value{
Returns a single log posterior probability or the prevalence from the
SEEIRR model.
}
\description{
Creates a new function which calculates the posterior probability of a set
of parameter values conditional on the Ct data using an SEEIRR model.
}
\seealso{
Other create posterior functions: 
\code{\link{create_post_func_seeirr_combined}()},
\code{\link{create_posterior_func_detectable}()},
\code{\link{create_posterior_func}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{create posterior functions}
