% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_posterior_func.R
\name{create_posterior_func_detectable}
\alias{create_posterior_func_detectable}
\title{Creates the posterior function used in the MCMC framework for detectable Ct values}
\usage{
create_posterior_func_detectable(
  parTab,
  data,
  PRIOR_FUNC = NULL,
  INCIDENCE_FUNC = NULL,
  solve_likelihood = TRUE,
  ...
)
}
\arguments{
\item{parTab}{A vector of model parameters with names corresponding to the parameter
control table.}

\item{data}{A tibble containing two columns: t for sample time and ct for the Ct values.}

\item{PRIOR_FUNC}{A function for the prior distribution. NULL by default.}

\item{INCIDENCE_FUNC}{A function that returns a vector of daily infection probabilities/
incidence. NULL by default.}

\item{solve_likelihood}{Solves the likelihood. TRUE by default.}
}
\value{
Returns a single log posterior probability.
}
\description{
Creates a new function which calculates the posterior probability of a set
of parameter values conditional on data that include only individuals with
detectable PCR values on the sample collection day.
}
\seealso{
Other create posterior functions: 
\code{\link{create_post_func_seeirr_combined}()},
\code{\link{create_post_func_seeirr}()},
\code{\link{create_posterior_func}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{create posterior functions}
