% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{detectable_SEIRModel}
\alias{detectable_SEIRModel}
\title{Detectable SEIR model}
\usage{
detectable_SEIRModel(pars, times)
}
\arguments{
\item{pars}{Vector of labeled SEIR model parameters.}

\item{times}{Vector of labeled time points for incidence calculation.}
}
\description{
This function solves the SEIR ODEs for a
series of time points and returns a vector
of detectable prevalence values.
}
\examples{
FIX ME

}
\seealso{
Other incidence functions: 
\code{\link{SEEIRRModel}()},
\code{\link{detectable_SEEIRRModel}()},
\code{\link{exponential_growth_model}()},
\code{\link{gaussian_process_model}()},
\code{\link{reverse_gp_model}()},
\code{\link{solveSEEIRRModel_rlsoda_wrapper}()},
\code{\link{solveSEEIRRModel_rlsoda}()},
\code{\link{solveSEIRModel_lsoda_wrapper}()},
\code{\link{solveSEIRModel_lsoda}()},
\code{\link{solveSEIRModel_rlsoda_wrapper}()},
\code{\link{solveSEIRModel_rlsoda}()},
\code{\link{solveSEIRswitch_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{incidence functions}
