% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ct_data}
\alias{example_ct_data}
\title{Example Ct value data}
\format{
A data frame with 15984 rows and 2 variables:
\describe{
\item{t}{Time in days that the sample was taken, relative to the start time of the simulation (simulation start time on t=0)}
\item{ct}{Ct value of the sample}
}
}
\usage{
data(example_ct_data)
}
\description{
Observed Ct values simulated from an SEIR model, with 2000 individuals sampled at random every 14 days between days 55 and 153 of the simulation. Each entry is one anonymised individual sample.
}
\seealso{
Other example_data: 
\code{\link{example_gp_partab}},
\code{\link{example_seir_incidence}},
\code{\link{example_seir_partab}}
}
\concept{example_data}
\keyword{datasets}
