% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_epi_data}
\alias{example_epi_data}
\title{Example epidemic data}
\format{
A data frame with 50000 rows and 5 variables:
\describe{
\item{Country}{Country associated with case numbers.}
\item{Region}{Region associated with case numbers.}
\item{Confirmed}{Confirmed case count (Cumulative)}
\item{Deceased}{Deceased case count (Cumulative)}
\item{Date}{Date of case reporting}
}
}
\usage{
data(example_epi_data)
}
\description{
Example of epidemic data pulled from a Covid-19 WHO dataset (link:)
to view epidemic trends with the virosolver data visualization
page. By default, the app uses this data to populate the
epi data visualization tab.
}
\seealso{
Other example_data: 
\code{\link{example_ct_data}},
\code{\link{example_gp_partab}},
\code{\link{example_seir_incidence}},
\code{\link{example_seir_partab}},
\code{\link{vignette_data1}},
\code{\link{vignette_data2}}
}
\concept{example_data}
\keyword{datasets}
