% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gp_partab}
\alias{example_gp_partab}
\title{Example Gaussian Process prior parameter table}
\format{
A data frame with 384 rows and 8 variables:
\describe{
\item{values}{numeric values of the parameters}
\item{names}{string names of the model parameters}
\item{fixed}{binary values indicating if the parameter should be fixed (1) or estimated (0) during the MCMC procedure}
\item{lower_bound}{lower numeric bound for the parameter during fitting (lower uniform prior bound)}
\item{upper_bound}{upper numeric bound for the parameter during fitting (upper uniform prior bound)}
\item{steps}{value between 0 and 1, giving the initial step size in the MCMC proposals. Note that these are adapted automatically}
\item{lower_start}{can be used to set the lower allowable random starting value for the MCMC}
\item{upper_start}{can be used to set the upper allowable random starting value for the MCMC}
}
}
\usage{
data(example_gp_partab)
}
\description{
Example of the parameter table used for to fit the Gaussian Process prior model with \code{virosolver}.
This data frame is used to control everything related to the model parameters, including values, names, fixed/estimates, and uniform prior bounds.
For this parameter table, we have entries for the GP prior parameters (the latent states for each day "prob", the hyperparameters for the exponential kernel function, rho and nu), as well as
entries for the viral kinetics/Ct model (peak Ct value, waning duration, time to peak, variation about the model Ct curve etc).
}
\seealso{
Other example_data: 
\code{\link{example_ct_data}},
\code{\link{example_epi_data}},
\code{\link{example_seir_incidence}},
\code{\link{example_seir_partab}},
\code{\link{vignette_data1}},
\code{\link{vignette_data2}}
}
\concept{example_data}
\keyword{datasets}
