% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_seir_incidence}
\alias{example_seir_incidence}
\title{True simulated SEIR incidence}
\format{
A tible with 251 rows and 2 columns
\describe{
\item{t}{Time in days}
\item{prob_infection}{Per capita incidence}
}
}
\usage{
data(example_seir_incidence)
}
\description{
The true SEIR incidence curve used in the case study simulation
}
\seealso{
Other example_data: 
\code{\link{example_ct_data}},
\code{\link{example_gp_partab}},
\code{\link{example_seir_partab}}
}
\concept{example_data}
\keyword{datasets}
