% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_seir_partab}
\alias{example_seir_partab}
\title{Example SEIR model parameter table}
\format{
A data frame with 20 rows and 8 variables:
\describe{
\item{values}{numeric values of the parameters}
\item{names}{string names of the model parameters}
\item{fixed}{binary values indicating if the parameter should be fixed (1) or estimated (0) during the MCMC procedure}
\item{lower_bound}{lower numeric bound for the parameter during fitting (lower uniform prior bound)}
\item{upper_bound}{upper numeric bound for the parameter during fitting (upper uniform prior bound)}
\item{steps}{value between 0 and 1, giving the initial step size in the MCMC proposals. Note that these are adapted automatically}
\item{lower_start}{can be used to set the lower allowable random starting value for the MCMC}
\item{upper_start}{can be used to set the upper allowable random starting value for the MCMC}
}
}
\usage{
data(example_seir_partab)
}
\description{
Example of the parameter table used for to simulate from an SEIR model with \code{virosolver}.
This data frame is used to control everything related to the model parameters, including values, names, fixed/estimates, and uniform prior bounds.
For this parameter table, we have entries for the SEIR model parameters (R0, infectious period, incubation period, seed time and seed size), as well as
entries for the viral kinetics/Ct model (peak Ct value, waning duration, time to peak, variation about the model Ct curve etc).
}
\seealso{
Other example_data: 
\code{\link{example_ct_data}},
\code{\link{example_gp_partab}},
\code{\link{example_seir_incidence}}
}
\concept{example_data}
\keyword{datasets}
