% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{exponential_growth_model}
\alias{exponential_growth_model}
\title{Exponential growth model}
\usage{
exponential_growth_model(pars, times)
}
\arguments{
\item{pars}{Vector of exponential growth model parameters.}

\item{times}{Vector of time points for incidence calculation.}
}
\value{
Returns incidence for given time points
}
\description{
This function returns incidence for given times
using the exponential growth model.
}
\examples{
FIX ME

}
\seealso{
Other incidence functions: 
\code{\link{SEEIRRModel}()},
\code{\link{detectable_SEEIRRModel}()},
\code{\link{detectable_SEIRModel}()},
\code{\link{gaussian_process_model}()},
\code{\link{reverse_gp_model}()},
\code{\link{solveSEEIRRModel_rlsoda_wrapper}()},
\code{\link{solveSEEIRRModel_rlsoda}()},
\code{\link{solveSEIRModel_lsoda_wrapper}()},
\code{\link{solveSEIRModel_lsoda}()},
\code{\link{solveSEIRModel_rlsoda_wrapper}()},
\code{\link{solveSEIRModel_rlsoda}()},
\code{\link{solveSEIRswitch_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{incidence functions}
