% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{generate_viable_start_pars}
\alias{generate_viable_start_pars}
\title{Function to generate random starting parameter values that
return a finite likelihood.}
\usage{
generate_viable_start_pars(
  parTab,
  obs_dat,
  CREATE_POSTERIOR_FUNC,
  INCIDENCE_FUNC,
  PRIOR_FUNC,
  use_pos = FALSE,
  ...
)
}
\arguments{
\item{obs_dat}{Dataframe containing observed Ct values.}

\item{CREATE_POSTERIOR_FUNC}{Creates the posterior function
used in the MCMC framework for detectable Ct values.}

\item{INCIDENCE_FUNC}{Function that expects a vector of named parameters
and returns a vector of daily incidence.}

\item{PRIOR_FUNC}{Function that returns the log prior probability for a
given vector of parameter values given the prior means and sds.}

\item{use_pos}{Boolean variable indicating if only positive Ct values
i.e, those below a specified threshold,  should be used. Set to FALSE
by default.}

\item{partab}{Dataframe containing parameters corresponding
to underlying incidence model (e.g. SEIR, SEEIRR).}
}
\value{
Dataframe containing random starting parameter values that return
a finite likelihood.
}
\description{
This function is used to generate random starting values
for the MCMC procedure.
}
\examples{
FIX ME

}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{mcmc help}
