% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihood_cpp_wrapper}
\alias{likelihood_cpp_wrapper}
\title{Wrapper function for the likelihood code, which is written in C.}
\usage{
likelihood_cpp_wrapper(
  obs_dat,
  ages,
  times,
  pars,
  prob_infection,
  pos_only = FALSE,
  undetectable_counts = NULL
)
}
\arguments{
\item{obs_dat}{Tibble of observed data. Contains two columns, t for time of sample
collection and ct for the Ct values.}

\item{ages}{Vector of times since infection.}

\item{times}{Vector of sample collection times in the data.}

\item{pars}{Model parameters.}

\item{prob_infection}{Probability of infection.}

\item{pos_only}{only use detectable Ct values? FALSE by default.}

\item{undetectable_counts}{Number of undetectable Ct samples. NULL by default.}
}
\value{
Returns the likelihood.
}
\description{
This function returns the likelihood of observing a particular set of
Ct values conditional on the underlying model parameters.
}
\examples{
FIX ME

}
\seealso{
Other likelihood functions: 
\code{\link{likelihood_R}()},
\code{\link{likelihood_detectable}()},
\code{\link{likelihood_pos_R}()},
\code{\link{p_a}()},
\code{\link{prob_detectable_curve}()},
\code{\link{prop_detectable_single}()},
\code{\link{prop_detectable}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{likelihood functions}
