% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihood_pos_R}
\alias{likelihood_pos_R}
\title{Likelihood for using proportion detectable only (R version)}
\usage{
likelihood_pos_R(obs_dat, ages, pars, prob_infection)
}
\arguments{
\item{obs_dat}{Tibble of observed data. Contains two columns, t for time of sample
collection and ct for the Ct values.}

\item{ages}{Vector of times since infection.}

\item{pars}{Model parameters.}

\item{prob_infection}{Probability of infection.}
}
\value{
Returns the likelihood.
}
\description{
This function is written in R. It returns the likelihood of observing a particular set of Ct
values conditional on the underlying model parameters using only the proportion
of Ct samples that are detectable.Despite being slower than the corresponding likelihood
function written in C, it more closely matches the written equations in the paper and
is retained for didactic purposes.
}
\examples{
FIX ME

}
\seealso{
Other likelihood functions: 
\code{\link{likelihood_R}()},
\code{\link{likelihood_cpp_wrapper}()},
\code{\link{likelihood_detectable}()},
\code{\link{p_a}()},
\code{\link{prob_detectable_curve}()},
\code{\link{prop_detectable_single}()},
\code{\link{prop_detectable}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{likelihood functions}
