% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{p_a}
\alias{p_a}
\title{Function that gives the probability of observing Ct value x given age
(days after infection) a and the viral kinetics curve.}
\usage{
p_a(x, a, pars, viral_loads, sd_mod)
}
\arguments{
\item{x}{Observed Ct value.}

\item{a}{Days after infection.}

\item{pars}{Model parameters.}

\item{viral_loads}{Observed viral loads.}

\item{sd_mod}{Standard deviation of the viral kinetics curve.}
}
\value{
Returns a vector of probabilities.
}
\description{
This is the Gumbel probability density function normalized to the observable values.
}
\seealso{
Other likelihood functions: 
\code{\link{likelihood_R}()},
\code{\link{likelihood_cpp_wrapper}()},
\code{\link{likelihood_detectable}()},
\code{\link{likelihood_pos_R}()},
\code{\link{prob_detectable_curve}()},
\code{\link{prop_detectable_single}()},
\code{\link{prop_detectable}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{likelihood functions}
