% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_posterior_density}
\alias{plot_posterior_density}
\title{Plot posterior density}
\usage{
plot_posterior_density(
  chain,
  var_name,
  parTab,
  prior_mean,
  prior_sd,
  prior_dist = "normal",
  real_data = FALSE
)
}
\arguments{
\item{chain}{}

\item{var_name}{Character string denoting variable name from parameter table.}

\item{parTab}{Dataframe specifying prior distribution for model parameter.}

\item{prior_mean}{Mean of the prior distribution for a particular parameter.}

\item{prior_sd}{Standard deviation of the prior distribution for a particular parameter.}

\item{real_data}{Plot true parameter values to simulate the data. Defaults to FALSE.}
}
\value{
ggplot
}
\description{
Plot posterior density from model output
}
\examples{
data(example_seir_partab)
chains <- load_mcmc_chains(location="mcmc_chains/readme_multiple_cross_section/",
parTab=par_tab,
burnin=mcmc_pars["adaptive_period"],
chainNo=FALSE,
unfixed=FALSE,
multi=FALSE)

var_name <- "tshift"
prior_mean <- subset(example_seir_partab, names == "tshift")$values 
prior_sd <- 0.2 
plot_posterior_density <- function(chain, var_name, parTab, prior_mean, prior_sd, real_data=FALSE)

}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{plot}
