% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prob_infection}
\alias{plot_prob_infection}
\title{Plot probability of infection}
\usage{
plot_prob_infection(
  chain,
  nsamps,
  INCIDENCE_FUNC,
  solve_times,
  obs_dat = NULL,
  true_prob_infection = NULL,
  tshift = 0,
  smooth = FALSE
)
}
\arguments{
\item{chain}{A dataframe containing the MCMC samples}

\item{nsamps}{Number of samples}

\item{INCIDENCE_FUNC}{A pointer to the Gaussian process model}

\item{solve_times}{Vector indicating the time over which the model is solved}

\item{obs_dat}{A dataframe containing observed Ct values and time of sample
collection. NULL by default.}

\item{true_prob_infection}{A dataframe from simulated data with two columns,
one for time and the other is the true probability of infection.
NULL by default.}

\item{tshift}{Shift the solve times? Numeric, set to 0 by default}

\item{smooth}{Smooth the model estimates for plotting? FALSE by default.}
}
\value{
Return a list containing three things: 1. A dataframe of model predictions containing
time, probability of infection, and sample number;
2. A dataframe containing the maximum posterior probability of infection and time;
3. A ggplot showing the probabilities of infection
}
\description{
Plot probabilities of infection from compartmental model. Returns the probabilities and the plot.
}
\examples{
data(example_seir_incidence)
predictions <- plot_prob_infection(chain_comb, 
                                 nsamps=100, 
                                 INCIDENCE_FUNC=incidence_function,
                                 solve_times=0:max(ct_data_use$t),
                                 obs_dat=ct_data_use,
                                 true_prob_infection=example_seir_incidence)
p_incidence_prediction <- predictions$plot + scale_x_continuous(limits=c(0,200))
p_incidence_prediction

}
\seealso{
Other plots: 
\code{\link{plot_distribution_fits}()},
\code{\link{predicted_distribution_fits}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{plots}
