% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{predicted_distribution_fits}
\alias{predicted_distribution_fits}
\title{Predicted distribution fits}
\usage{
predicted_distribution_fits(chain, MODEL_FUNC, nsamps = 100)
}
\arguments{
\item{chain}{A dataframe containing the MCMC samples}

\item{MODEL_FUNC}{Function that expects a vector of model parameters with names
corresponding to the parameter control table and returns a single log posterior probability}

\item{nsamps}{Number of samples. Defaults to 100.}
}
\value{
Returns a dataframe containing the predictions from the posterior distribution.
}
\description{
Obtain predicted Ct distribution fits from model (posterior_dat)
}
\examples{
data(example_ct_data)
data(example_seir_partab)

\dontrun{

MODEL_FUNC <- create_posterior_func(parTab=example_seir_partab,
                                     data=example_ct_data,
                                     PRIOR_FUNC=prior_func_seir,
                                     INCIDENCE_FUNC=incidence_function,
                                     use_pos=FALSE) 
                                     
posterior_dat <- predicted_distribution_fits(chain, MODEL_FUNC, nsamps=100)                                   
head(posterior_dat)
}

}
\seealso{
Other plots: 
\code{\link{plot_distribution_fits}()},
\code{\link{plot_prob_infection}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{plots}
