% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{prop_detectable}
\alias{prop_detectable}
\title{Proportion still detectable (range of times)}
\usage{
prop_detectable(ages, pars, viral_loads)
}
\arguments{
\item{ages}{Vector of times since infection.}

\item{pars}{Model parameters.}

\item{viral_loads}{Observed viral loads.}
}
\value{
Returns a vector of probabilities.
}
\description{
Gives the proportion of infections still detectable over a range of times
since infection (referred to as 'ages')
}
\examples{
data(example_gp_partab)
pars <- example_gp_partab$values
names(pars) <- example_gp_partab$names
test_ages <- seq(1,50,by=1)
cts <- viral_load_func(pars, test_ages)
prop_detect <- prop_detectable(test_ages,pars, cts)
}
\seealso{
Other likelihood functions: 
\code{\link{likelihood_R}()},
\code{\link{likelihood_cpp_wrapper}()},
\code{\link{likelihood_detectable}()},
\code{\link{likelihood_pos_R}()},
\code{\link{p_a}()},
\code{\link{prob_detectable_curve}()},
\code{\link{prop_detectable_single}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{likelihood functions}
