% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{prop_detectable_single}
\alias{prop_detectable_single}
\title{Probability of having a detectable Ct for a given time since infection}
\usage{
prop_detectable_single(a, pars, viral_loads, sd_mod)
}
\description{
Probability of having a detectable Ct for a given time since infection
}
