% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{reverse_gp_model}
\alias{reverse_gp_model}
\title{Reverse Gaussian process model}
\usage{
reverse_gp_model(desired_probs, pars, times)
}
\arguments{
\item{desired_probs}{Vector of probabilities for GP transformation.}

\item{pars}{Exponential growth model parameters.}

\item{times}{Time points for incidence calculation.}
}
\value{
Returns daily infection probabilities in the format
expected by the Gaussian process model.
}
\description{
Given a vector of daily infection probabilities,
this function converts this vector to the format expected
by the Gaussian process model.
}
\examples{
FIX ME

}
\seealso{
Other incidence functions: 
\code{\link{SEEIRRModel}()},
\code{\link{detectable_SEEIRRModel}()},
\code{\link{detectable_SEIRModel}()},
\code{\link{exponential_growth_model}()},
\code{\link{gaussian_process_model}()},
\code{\link{solveSEEIRRModel_rlsoda_wrapper}()},
\code{\link{solveSEEIRRModel_rlsoda}()},
\code{\link{solveSEIRModel_lsoda_wrapper}()},
\code{\link{solveSEIRModel_lsoda}()},
\code{\link{solveSEIRModel_rlsoda_wrapper}()},
\code{\link{solveSEIRModel_rlsoda}()},
\code{\link{solveSEIRswitch_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{incidence functions}
