% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{SEEIRRModel}
\alias{SEEIRRModel}
\title{SEEIRR Model}
\usage{
SEEIRRModel(pars, times)
}
\arguments{
\item{pars}{Vector of labeled SEEIRR model parameters.}

\item{times}{Vector of times points.}
}
\value{
Returns dataframe of variables solved for using SEEIRR ODEs.
}
\description{
This function returns a dataframe containing the values solved for
using the SEEIRR differential equation set.
}
\examples{
FIX ME

}
\seealso{
Other incidence functions: 
\code{\link{detectable_SEEIRRModel}()},
\code{\link{detectable_SEIRModel}()},
\code{\link{exponential_growth_model}()},
\code{\link{gaussian_process_model}()},
\code{\link{reverse_gp_model}()},
\code{\link{solveSEEIRRModel_rlsoda_wrapper}()},
\code{\link{solveSEEIRRModel_rlsoda}()},
\code{\link{solveSEIRModel_lsoda_wrapper}()},
\code{\link{solveSEIRModel_lsoda}()},
\code{\link{solveSEIRModel_rlsoda_wrapper}()},
\code{\link{solveSEIRModel_rlsoda}()},
\code{\link{solveSEIRswitch_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{incidence functions}
