% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_detectable_wrapper}
\alias{simulate_detectable_wrapper}
\title{Simulate binary PCR results (detectables) for the line list dataset.}
\usage{
simulate_detectable_wrapper(linelist, pars)
}
\arguments{
\item{linelist}{the line list for observed individuals}

\item{pars}{vector of named parameters for the detectability curve model}
}
\value{
A tibble with the line list data and the PCR status at the time of sampled
}
\description{
Simulate binary PCR results (detectables) for the line list dataset.
}
\seealso{
Other viral load functions: 
\code{\link{pred_dist_wrapper}()},
\code{\link{simulate_viral_loads_example_symptoms}()},
\code{\link{simulate_viral_loads_example}()},
\code{\link{viral_load_func}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{viral load functions}
