% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_infection_times}
\alias{simulate_infection_times}
\title{Simulate infection times}
\usage{
simulate_infection_times(n, prob_infection, overall_prob = NULL)
}
\arguments{
\item{n}{Sample size. Must be an integer.}

\item{prob_infection}{A vector containing probabilities of infection.}

\item{overall_prob}{The overall probability of infection. NULL by default.}
}
\value{
Returns a vector of infection times.
}
\description{
Infection times are simulated using the probability of infection and sample size n.
}
\examples{
FIX ME

}
\seealso{
Other simulation functions: 
\code{\link{simulate_observations_wrapper}()},
\code{\link{simulate_reporting}()},
\code{\link{simulate_seir_process}()},
\code{\link{simulate_viral_loads_wrapper}()},
\code{\link{simulate_viral_loads}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{simulation functions}
