% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_observations_wrapper}
\alias{simulate_observations_wrapper}
\title{Simulate full line list data}
\usage{
simulate_observations_wrapper(
  incidence,
  times,
  symp_frac = 0.35,
  population_n = length(incidence),
  incu_period_par1 = 1.621,
  incu_period_par2 = 0.418,
  conf_delay_par1 = 5,
  conf_delay_par2 = 2
)
}
\description{
Takes a vector of infection incidence (absolute numbers) and returns a tibble with line list data for all
individuals in the population. Infected individuals are assigned symptom onset, an incubation periods (if applicable)
and confirmation delays from log normal and gamma distributions respectively.
}
