% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_reporting}
\alias{simulate_reporting}
\title{Subset line list data by testing strategy. Options:
\enumerate{
\item Sample a random fraction of the population if the only argument is frac_report
\item Sample some random fraction of the population at a subset of time points, specified by timevarying_prob
\item Observe symptomatic individuals with some fixed probability, frac_report if symptomatic is TRUE
\item Observe symptomatic individuals with some time-varying probability, timevarying_prob, if symptomatic is TRUE
INPUTS:
\enumerate{
\item individuals: the full line list from the simulation, returned by virosolver::simulate_observations_wrapper
\item solve_times: vector of times at which individuals can be reported
\item frac_report: the overall fraction/probability of individuals who are reported
\item timevarying_prob: a tibble with variables t and prob. This gives the probability of being reported on day t
\item symptomatic: if TRUE, then individuals are reported after developing symptoms. If FALSE, then we take a random cross-section
OUTPUTS:
\item A tibble with line list data for individuals who were observed
\item A plot of incidence for both observed individuals and the entire simulated population
\item Plot growth rate of cases/infections in the entire population and observed population
}
}}
\usage{
simulate_reporting(
  individuals,
  solve_times,
  frac_report = 1,
  timevarying_prob = NULL,
  symptomatic = FALSE
)
}
\description{
Subset line list data by testing strategy. Options:
\enumerate{
\item Sample a random fraction of the population if the only argument is frac_report
\item Sample some random fraction of the population at a subset of time points, specified by timevarying_prob
\item Observe symptomatic individuals with some fixed probability, frac_report if symptomatic is TRUE
\item Observe symptomatic individuals with some time-varying probability, timevarying_prob, if symptomatic is TRUE
INPUTS:
\enumerate{
\item individuals: the full line list from the simulation, returned by virosolver::simulate_observations_wrapper
\item solve_times: vector of times at which individuals can be reported
\item frac_report: the overall fraction/probability of individuals who are reported
\item timevarying_prob: a tibble with variables t and prob. This gives the probability of being reported on day t
\item symptomatic: if TRUE, then individuals are reported after developing symptoms. If FALSE, then we take a random cross-section
OUTPUTS:
\item A tibble with line list data for individuals who were observed
\item A plot of incidence for both observed individuals and the entire simulated population
\item Plot growth rate of cases/infections in the entire population and observed population
}
}
}
