% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_epidemics.R
\name{simulate_seir_process}
\alias{simulate_seir_process}
\title{Simulate SEIR model}
\usage{
simulate_seir_process(pars, times, N = 1e+05, switch_model = FALSE)
}
\arguments{
\item{pars}{SEIR model parameters.}

\item{times}{Times over which the model is solved.}

\item{N}{Population size. Defaults to 100000.}
}
\value{
Returns a list of 7 things:
\enumerate{
\item Plot of all SEIR compartments over time
\item Plot of incidence and prevalence over time
\item Solution of ordinary differential equation
\item Absolute incidence per time point (raw incidence)
\item Per capita incidence per time point
\item Per capita prevalence (compartments E+I) per time point
\item Overall probability of infection
}
}
\description{
Simulates a deterministic SEIR model from model parameters, times to solve over, and
population size.
}
\seealso{
Other simulation functions: 
\code{\link{simulate_infection_times}()},
\code{\link{simulate_observations_wrapper}()},
\code{\link{simulate_reporting}()},
\code{\link{simulate_seir_wrapper}()},
\code{\link{simulate_viral_loads_wrapper}()},
\code{\link{simulate_viral_loads}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{simulation functions}
