% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_epidemics.R
\name{simulate_seir_wrapper}
\alias{simulate_seir_wrapper}
\title{Wrapper for SEIR model simulation, deterministic or stochastic}
\usage{
simulate_seir_wrapper(
  N = 1e+05,
  times,
  pars,
  version = "ode",
  switch_model = FALSE,
  beta_smooth = 0.8
)
}
\value{
Returns a list of 6 things:
1. Data frame of the SEIR solution
2. Vector of absolute incidence per time point
3. Overall probability of infection
4. Plot of incidence over time
5. Average growth rates over different time periods
6. Plot of average growth rates over different time periods
}
\description{
Simulates a deterministic or stochastic SEIR model from model parameters
}
\details{
@param N Population size. Defaults to 100000.
@param times Vector of times to solve model over
@param pars Named vector of SEIR model parameters
@param version String either "ode" for the deterministic version or "odin" for stochastic
@param switch_model If TRUE, uses the SEIR model with 2 switch points in transmission intensity
@param beta_smooth spar parameter fed to smooth.spline, used to smooth the time-varying beta
}
\seealso{
Other simulation functions: 
\code{\link{simulate_infection_times}()},
\code{\link{simulate_observations_wrapper}()},
\code{\link{simulate_reporting}()},
\code{\link{simulate_seir_process}()},
\code{\link{simulate_viral_loads_wrapper}()},
\code{\link{simulate_viral_loads}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{simulation functions}
