% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_viral_loads}
\alias{simulate_viral_loads}
\title{Simulate viral loads}
\usage{
simulate_viral_loads(
  infection_times,
  solve_times,
  kinetics_pars,
  vl_func_use = viral_load_func,
  convert_vl = FALSE,
  add_noise = NULL
)
}
\arguments{
\item{infection_times}{Vector containing infection times.}

\item{solve_times}{Vector of times at which individuals can be reported.}

\item{kinetics_pars}{Vector of named parameters for the viral kinetics model.}

\item{vl_func_use}{Determines which viral load function to use.
The default is viral_load_func.}

\item{convert_vl}{Convert to viral load value. FALSE by default.}

\item{add_noise}{Add noise to Ct values? NULL by default.}
}
\value{
Returns a tibble containing true and observed viral loads.
}
\description{
Takes infection times, the times over which to solve the model, and
the viral kinetics parameters and returns a tibble containing true and observed viral loads.
}
\seealso{
Other simulation functions: 
\code{\link{simulate_infection_times}()},
\code{\link{simulate_observations_wrapper}()},
\code{\link{simulate_reporting}()},
\code{\link{simulate_seir_process}()},
\code{\link{simulate_seir_wrapper}()},
\code{\link{simulate_viral_loads_wrapper}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{simulation functions}
