% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_viral_loads_example}
\alias{simulate_viral_loads_example}
\title{This function simulates N Ct values per age in ages
and returns a dataframe of Ct values and ages.}
\usage{
simulate_viral_loads_example(ages, kinetics_pars, N = 100)
}
\arguments{
\item{ages}{Vector of ages (days since infection).}

\item{kinetics_pars}{Vector of named parameters for the viral kinetics model.}

\item{N}{Number of observations to randomly generate per age;
Set to 100 by default.}
}
\value{
Dataframe of simulated Ct values and corresponding ages.
}
\description{
This function simulates N Ct values per age in ages
and returns a dataframe of Ct values and ages.
}
\seealso{
Other viral load functions: 
\code{\link{pred_dist_wrapper}()},
\code{\link{simulate_detectable_wrapper}()},
\code{\link{simulate_viral_loads_example_symptoms}()},
\code{\link{viral_load_func}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{viral load functions}
