% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_viral_loads_example_symptoms}
\alias{simulate_viral_loads_example_symptoms}
\title{Simulate N Ct values under symptom-based surveillance
and returns a dataframe of Ct values and times since infection.}
\usage{
simulate_viral_loads_example_symptoms(
  ages,
  kinetics_pars,
  incu_par1 = 1.621,
  incu_par2 = 0.418,
  sampling_par1 = 5,
  sampling_par2 = 1,
  N = 100,
  sampling_dist_use = extraDistr::rdgamma
)
}
\arguments{
\item{ages}{Vector of ages (days since infection).}

\item{kinetics_pars}{Vector of named parameters for the viral kinetics model.}

\item{incu_par1}{logMean of the log-normal incubation period distribution}

\item{incu_par2}{logSD of the log-normal incubation period distribution}

\item{sampling_par1}{shape of the discretized gamma sampling delay distribution}

\item{sampling_par2}{rate of the discretized gamma sampling delay distribution}

\item{N}{Number of observations to randomly generate overall;
Set to 100 by default.}
}
\value{
Dataframe of simulated Ct values and corresponding ages.
}
\description{
Simulate N Ct values under symptom-based surveillance
and returns a dataframe of Ct values and times since infection.
}
\seealso{
Other viral load functions: 
\code{\link{pred_dist_wrapper}()},
\code{\link{simulate_detectable_wrapper}()},
\code{\link{simulate_viral_loads_example}()},
\code{\link{viral_load_func}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{viral load functions}
