% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{solveSEIRModel_lsoda}
\alias{solveSEIRModel_lsoda}
\title{Solve SEIR Model with lsoda}
\usage{
solveSEIRModel_lsoda(ts, init, pars, compatible = FALSE)
}
\arguments{
\item{init}{Vector of labeled initial conditions for system of equations.}

\item{pars}{Vector of SEIR model parameters.}

\item{times}{Vector of time points.}
}
\value{
Returns matrix of class deSolve containing solutions for SEIR ODEs
for each time in ts.
}
\description{
This function solves the SEIR ODEs and returns
a matrix of class deSolve.
}
\examples{
FIX ME

}
\seealso{
Other incidence functions: 
\code{\link{SEEIRRModel}()},
\code{\link{detectable_SEEIRRModel}()},
\code{\link{detectable_SEIRModel}()},
\code{\link{exponential_growth_model}()},
\code{\link{gaussian_process_model}()},
\code{\link{reverse_gp_model}()},
\code{\link{solveSEEIRRModel_rlsoda_wrapper}()},
\code{\link{solveSEEIRRModel_rlsoda}()},
\code{\link{solveSEIRModel_lsoda_wrapper}()},
\code{\link{solveSEIRModel_rlsoda_wrapper}()},
\code{\link{solveSEIRModel_rlsoda}()},
\code{\link{solveSEIRswitch_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{incidence functions}
