% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{solveSEIRModel_rlsoda}
\alias{solveSEIRModel_rlsoda}
\title{Solve SEIR model with rlsoda}
\usage{
solveSEIRModel_rlsoda(ts, init, pars, compatible = FALSE)
}
\arguments{
\item{ts}{Vector of time points.}

\item{init}{Vector of initial conditions for system of equations.}

\item{pars}{Vector of labeled SEIR model parameters.}

\item{compatible}{A Boolean variable indicating if
results should be transposed in deSolve compatible format.
Set to FALSE by default.}
}
\value{
Returns matrix of class deSolve containing solutions for SEIR ODEs
for each time in ts.
}
\description{
This function solves the SEIR ODEs for a series of
time points and returns a matrix of the solved values. It uses rlsoda
which must be manually downloaded from GitHub.
}
\examples{
FIX ME

}
\seealso{
Other incidence functions: 
\code{\link{SEEIRRModel}()},
\code{\link{detectable_SEEIRRModel}()},
\code{\link{detectable_SEIRModel}()},
\code{\link{exponential_growth_model}()},
\code{\link{gaussian_process_model}()},
\code{\link{reverse_gp_model}()},
\code{\link{solveSEEIRRModel_rlsoda_wrapper}()},
\code{\link{solveSEEIRRModel_rlsoda}()},
\code{\link{solveSEIRModel_lsoda_wrapper}()},
\code{\link{solveSEIRModel_lsoda}()},
\code{\link{solveSEIRModel_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda_wrapper}()},
\code{\link{solveSEIRswitch_rlsoda}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{incidence functions}
