% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_data1}
\alias{vignette_data1}
\title{Simulated SEIR incidence for vignettes}
\format{
Numeric vector with length 151
}
\usage{
data(vignette_data1)
}
\description{
This is the raw incidence from the simulate_seir_process function. It is created in vignette 1 and used in vignette 2.
}
\seealso{
Other example_data: 
\code{\link{example_ct_data}},
\code{\link{example_epi_data}},
\code{\link{example_gp_partab}},
\code{\link{example_seir_incidence}},
\code{\link{example_seir_partab}},
\code{\link{vignette_data2}}
}
\concept{example_data}
\keyword{datasets}
