% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_data2}
\alias{vignette_data2}
\title{Simulated Ct values for vignettes}
\format{
A data frame with 22237 rows and and 4 columns
\describe{
\item{sampled_time}{Time of sample collection in days}
\item{Ct}{Ct values}
\item{date}{calendar date}
\item{Ct_round}{rounded Ct values}
}
}
\usage{
data(vignette_data2)
}
\description{
These are the simulated Ct values which are created in vignette 1 and used in vignette 2.
}
\seealso{
Other example_data: 
\code{\link{example_ct_data}},
\code{\link{example_epi_data}},
\code{\link{example_gp_partab}},
\code{\link{example_seir_incidence}},
\code{\link{example_seir_partab}},
\code{\link{vignette_data1}}
}
\concept{example_data}
\keyword{datasets}
