% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_load_functions.R
\name{viral_load_func}
\alias{viral_load_func}
\title{This function returns modal Ct values or viral load values, if
convert_vl flag is set to TRUE.}
\usage{
viral_load_func(pars, obs_t, convert_vl = FALSE, infection_time = 0)
}
\arguments{
\item{pars}{Vector of model parameter values.}

\item{obs_t}{Vector of times since infection. Referred to as "ages."}

\item{convert_vl}{Boolean flag used to indicate if Ct values should
be converted to viral load values. Set to FALSE by default.}

\item{infection_time}{Time of infection. Set to 0 by default.}
}
\value{
Returns modal Ct or viral load values.
}
\description{
This function returns modal Ct values or viral load values, if
convert_vl flag is set to TRUE.
}
\seealso{
Other viral load functions: 
\code{\link{pred_dist_wrapper}()},
\code{\link{simulate_detectable_wrapper}()},
\code{\link{simulate_viral_loads_example_symptoms}()},
\code{\link{simulate_viral_loads_example}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{viral load functions}
