% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihood_detectable}
\alias{likelihood_detectable}
\title{Likelihood for using proportion detectable only}
\usage{
likelihood_detectable(obs_dat, ages, pars, prob_infection)
}
\description{
Likelihood for using proportion detectable only
}
