% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{p_a}
\alias{p_a}
\title{Function to give probability of observing x given age a and the viral kinetics curve}
\usage{
p_a(x, a, pars, viral_loads, sd_mod)
}
\description{
Function to give probability of observing x given age a and the viral kinetics curve
}
