% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{prop_detectable}
\alias{prop_detectable}
\title{Probability of having a detectable Ct for a given time since infection}
\usage{
prop_detectable(a, pars, viral_loads)
}
\description{
Probability of having a detectable Ct for a given time since infection
}
