% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{prop_detectable_single}
\alias{prop_detectable_single}
\title{Proportion still detectable (single time)}
\usage{
prop_detectable_single(a, pars, viral_loads, sd_mod)
}
\arguments{
\item{a}{Days after infection.}

\item{pars}{Model parameters.}

\item{viral_loads}{Observed viral loads.}

\item{sd_mod}{Standard deviation of the viral kinetics curve.}
}
\value{
Returns a vector of probabilities.
}
\description{
Gives the proportion of infections still detectable for a single time since infection (a)
}
\seealso{
Other likelihood functions: 
\code{\link{likelihood_R}()},
\code{\link{likelihood_cpp_wrapper}()},
\code{\link{likelihood_detectable}()},
\code{\link{likelihood_pos_R}()},
\code{\link{p_a}()},
\code{\link{prob_detectable_curve}()},
\code{\link{prop_detectable}()}
}
\author{
James Hay, \email{jhay@hsph.harvard.edu}
}
\concept{likelihood functions}
