% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_functions.R
\name{reverse_gp_model}
\alias{reverse_gp_model}
\title{Given a vector of daily infection probabilities, converts this to the input expected by the Gaussian process model}
\usage{
reverse_gp_model(desired_probs, pars, times)
}
\description{
Given a vector of daily infection probabilities, converts this to the input expected by the Gaussian process model
}
