% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_observations_wrapper}
\alias{simulate_observations_wrapper}
\title{Simulate full line list data}
\usage{
simulate_observations_wrapper(
  incidence,
  times,
  symp_frac = 0.35,
  population_n = 1e+05,
  incu_period_par1 = 1.621,
  incu_period_par2 = 0.418,
  conf_delay_par1 = 5,
  conf_delay_par2 = 2,
  sampling_dist_use = extraDistr::rdgamma
)
}
\arguments{
\item{incidence}{A vector of infection incidence (absolute numbers).}

\item{times}{Calendar time (i.e. days since the start of the epidemic).}

\item{symp_frac}{Fraction of the population that is symptomatic. Defaults to 0.35.}

\item{population_n}{Size of the population. Defaults to 100,000. Must match the N used
in the simulate_seir_process function.}

\item{incu_period_par1}{The first parameter (meanLog) associated with the log-normal incubation period.
Defaults to 1.621.}

\item{incu_period_par2}{The second parameter (sdLog) associated with the log-normal incubation period.
Defaults to 0.418.}

\item{conf_delay_par1}{The first parameter (shape) associated with the discretized gamma confirmation delay.
Defaults to 5.}

\item{conf_delay_par2}{The second parameter (rate) associated with the discretized gamma confirmation delay.
Defaults to 2.}
}
\value{
A tibble with line list data for all individuals in the population.
}
\description{
Takes a vector of infection incidence (absolute numbers) and returns a tibble with line list data for all
individuals in the population. Infected individuals are assigned symptom onset, incubation
periods (if applicable) and confirmation delays from log normal and gamma distributions respectively.
}
\seealso{
Other simulation functions: 
\code{\link{simulate_infection_times}()},
\code{\link{simulate_reporting}()},
\code{\link{simulate_seir_process}()},
\code{\link{simulate_seir_wrapper}()},
\code{\link{simulate_viral_loads_wrapper}()},
\code{\link{simulate_viral_loads}()}
}
\author{
James Hay, \email{jameshay218@gmail.com}
}
\concept{simulation functions}
