% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_viral_loads_wrapper}
\alias{simulate_viral_loads_wrapper}
\title{Simulate observed Ct values for the line list dataset. NOTE this differs to virosolver::simulate_viral_loads,
as this function only solves the viral kinetics model for the observation time
INPUTS:
1. linelist: the line list for observed individuals
2. kinetics_pars: vector of named parameters for the viral kinetics model
OUTPUTS:
1. A tibble with the line list data and the viral load/ct/observed ct at the time of sampled}
\usage{
simulate_viral_loads_wrapper(linelist, kinetics_pars)
}
\description{
Simulate observed Ct values for the line list dataset. NOTE this differs to virosolver::simulate_viral_loads,
as this function only solves the viral kinetics model for the observation time
INPUTS:
1. linelist: the line list for observed individuals
2. kinetics_pars: vector of named parameters for the viral kinetics model
OUTPUTS:
1. A tibble with the line list data and the viral load/ct/observed ct at the time of sampled
}
